Move environment file
